# server/swagger.json
{"openapi":"3.0.0","info":{"title":"API Documentazione","version":"1.0.0","description":"Documentazione delle API per la gestione utenti.","contact":{"name":"Gianmario Iamoni","email":"gianmarioiamoni1@gmail.com"}},"servers":[{"url":"http://localhost:5000/api"}],"paths":{"/users":{"get":{"summary":"Ottieni tutti gli utenti","description":"Restituisce una lista di tutti gli utenti registrati, con supporto per la paginazione.","parameters":[{"in":"query","name":"page","schema":{"type":"integer","default":1},"description":"Numero della pagina da visualizzare."},{"in":"query","name":"limit","schema":{"type":"integer","default":10},"description":"Numero di utenti da mostrare per pagina."},{"in":"query","name":"fields","schema":{"type":"string","default":"nome cognome email"},"description":"Campi degli utenti da includere nella risposta, separati da spazi."}],"responses":{"200":{"description":"Lista degli utenti con metadati di paginazione.","content":{"application/json":{"schema":{"type":"object","properties":{"users":{"type":"array","items":{"type":"object","properties":{"id":{"type":"string"},"nome":{"type":"string"},"cognome":{"type":"string"},"email":{"type":"string"}}}},"total":{"type":"integer","description":"Numero totale di utenti disponibili."}}}}}}}},"post":{"summary":"Crea un nuovo utente","description":"Aggiunge un nuovo utente al database. L'email deve essere univoca.","requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["nome","cognome","email"],"properties":{"nome":{"type":"string","description":"Nome dell'utente."},"cognome":{"type":"string","description":"Cognome dell'utente."},"email":{"type":"string","description":"Email univoca dell'utente."},"dataNascita":{"type":"string","format":"date","description":"Data di nascita dell'utente (opzionale)."},"fotoProfilo":{"type":"string","description":"URL della foto profilo (opzionale)."}}}}}},"responses":{"201":{"description":"Utente creato con successo.","content":{"application/json":{"schema":{"type":"object","properties":{"id":{"type":"string"},"nome":{"type":"string"},"cognome":{"type":"string"},"email":{"type":"string"},"dataNascita":{"type":"string","format":"date"},"fotoProfilo":{"type":"string"}}}}}},"400":{"description":"Dati non validi o mancanti."},"409":{"description":"Email già in uso."}}}},"/total":{"get":{"summary":"Ottieni il numero totale di utenti","description":"Restituisce il numero totale di documenti nella collezione degli utenti.","tags":["Users"],"responses":{"200":{"description":"Numero totale di utenti restituito con successo.","content":{"application/json":{"schema":{"type":"integer","example":150}}}},"500":{"description":"Errore interno del server."}}}},"/users/{id}":{"get":{"summary":"Ottieni un utente per ID","description":"Restituisce i dettagli di un singolo utente specificato dall'ID. L'ID deve essere un ObjectId valido.","parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string"},"description":"ID univoco dell'utente."}],"responses":{"200":{"description":"Dettagli dell'utente.","content":{"application/json":{"schema":{"type":"object","properties":{"id":{"type":"string"},"nome":{"type":"string"},"cognome":{"type":"string"},"email":{"type":"string"}}}}}},"400":{"description":"Formato dell'ID non valido."},"404":{"description":"Utente non trovato."}}}}},"components":{},"tags":[]}